refactor: MainView 구조 개선 및 사용자 정보 외부 주입
